+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [NEXT_RELEASE]
+### Added
+- Support for non-web process types
+
 ## [0.2.2] - 2017-03-16
 ### Changed
 - Upgrade slugbuilder and slugrunner